+++ examples/workerd-cli/e2e/basic.test.ts
@@ -1,5 +1,6 @@
 import test from "node:test";
 import childProcess from "node:child_process";
+import { createProcessHelper } from "./helper";
 
 test("basic", async () => {
   using proc = childProcess.spawn("pnpm", ["-s", "cli"]);
@@ -14,43 +15,3 @@ test("basic", async () => {
     out.includes("@hiogawa/vite-environment-examples-workerd-cli"),
   );
 });
-
-function createProcessHelper(
-  proc: childProcess.ChildProcessWithoutNullStreams,
-) {
-  let stdout = "";
-  const listeners = new Set<() => void>();
-  proc.stdout.on("data", (data) => {
-    stdout += String(data);
-    for (const f of listeners) {
-      f();
-    }
-  });
-
-  async function waitFor(predicate: (stdout: string) => boolean) {
-    const promise = new Promise<void>((resolve) => {
-      const listener = () => {
-        if (predicate(stdout)) {
-          resolve();
-          listeners.delete(listener);
-        }
-      };
-      listeners.add(listener);
-    });
-    const timeout = sleep(5000).then(() => {
-      throw new Error("waitFor timeout", { cause: stdout });
-    });
-    return Promise.race([promise, timeout]);
-  }
-
-  return {
-    get stdout() {
-      return stdout;
-    },
-    waitFor,
-  };
-}
-
-function sleep(ms: number) {
-  return new Promise((resolve) => setTimeout(resolve, ms).unref());
-}
